backend/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
sqlalchemy
alembic
python-dotenv>=1.0.0
psycopg2-binary
dotenv
bcrypt
pyjwt